requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
jupyter>=1.0.0
flask>=2.0.0
pillow>=9.0.0
tensorflow>=2.10.0
opencv-python>=4.5.0 